requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
flask-session
python-dotenv
flask-session
redis